Make Project as SinglePage
