+++ src/pages/ListPage.jsx
@@ -2,12 +2,10 @@ import Footer from "components/common/Footer";
 import H2Box from "components/common/H2Box";
 import ItemList from "components/list/ItemList";
 import ListFilter from "components/list/ListFilter";
-import Header from "components/main/Header";
 import { PageLayout } from "lib/styles/utilStyles";
 import React, { useState } from "react";
 import { categoryList } from "lib/categoryList";
 import { useParams } from "react-router-dom";
-import NavBar from "components/navigator/NavBar";
 
 function ListPage() {
 	const { catNum } = useParams();
@@ -18,8 +16,6 @@ function ListPage() {
 	};
 	return (
 		<>
-			<Header getToggled={getToggled} />
-			<NavBar toggled={toggled} getToggled={getToggled} />
 			<PageLayout>
 				<H2Box>{cat}</H2Box>
 				<ListFilter></ListFilter>
