+++ src/pages/MainPage.jsx
@@ -1,21 +1,13 @@
-import React, { useState } from "react";
-import Header from "../components/main/Header";
+import React from "react";
 import styled from "styled-components";
 import Search from "../components/search/Search";
 import Category from "../components/main/Category";
 import MainList from "components/main/MainList";
 import Footer from "components/common/Footer";
-import NavBar from "components/navigator/NavBar";
 
 function MainPage() {
-	const [toggled, setToggled] = useState(false);
-	const getToggled = (value) => {
-		setToggled(value);
-	};
 	return (
 		<MainLayout>
-			<Header getToggled={getToggled} />
-			<NavBar toggled={toggled} getToggled={getToggled} />
 			<Search />
 			<Category />
 			<MainList />
